fix: resolve all compiler warnings in Release build
- Remove unused variables: x_acc (coding_units.cpp), stride4 and t2
  (ht_block_encoding_avx512.cpp)
- Check fread return value (main_jpip_benchmark.cpp)
- Add explicit cast for int-to-size_t sign conversion (j2kmarkers.cpp)
- Fix misleading indentation: split same-line if/assignment
  (jpip_index_check/main.cpp)
- Add ptrdiff_t cast for size_t-to-difference_type conversion
  (jpip_parser_check/main.cpp)

+++ source/core/codestream/j2kmarkers.cpp
@@ -1440,7 +1440,7 @@ TLM_marker::TLM_marker(uint8_t ztlm, const std::vector<uint16_t> &tile_indices,
 void TLM_marker::write(j2c_dst_memory &buf) const {
   uint8_t ST = (Stlm >> 4) & 0x03;
   uint8_t SP = ((Stlm >> 4) & 0x0C) >> 2;
-  size_t entry_size = (ST == 0 ? 0 : (ST == 1 ? 1 : 2)) + (SP == 0 ? 2 : 4);
+  size_t entry_size = static_cast<size_t>((ST == 0 ? 0 : (ST == 1 ? 1 : 2)) + (SP == 0 ? 2 : 4));
   uint16_t Ltlm = static_cast<uint16_t>(4 + Ptlm.size() * entry_size);
   buf.put_word(_TLM);
   buf.put_word(Ltlm);
